filter/timing: add realtime mode
The realtime mode is a high-throughput variant of the keep mode. It
is less precise, but is more efficient and able to handle higher
throughput.

If there are significant time gaps in the input packets, the realtime
mode may send them earlier if both packets end up in the same batch.
However, packets are guaranteed to never be sent later than the
specified drift. In cases where the timer isn't able to keep up for any
reason, it aborts the execution.

The default batch size of 128 is an empirical value that has been tested
to perform well with input data of up to ~500k pps.