output.pcap: detect write errors when closing output
Problem:
libpcap brain-dead API does not provide error codes for file operations
and accepts packets indefinitelly even if the underlying file cannot be
written to, e.g. bacause of ENOSPC.

Workaround:
fopen() file in our code and check ferror() when closing file.
Lots of packets can be lost before the file is closed, but at least it
is detected at the very end, which is an improvement over undetectable
packet loss.

This commit does not have measurable performance impact.